fix: correct source binding when calling `add_aggregates`
Fixes problem adding aggregates to a nested relationship. Not 100% certain if adding 1 to the current binding is the correct fix.

My intuition of why this previously worked in most cases is because the key attributes often overlap between a join table and the table aggregates are being added too. Alternatively this fix is introducing new bugs.

Needs code review.